# api/SWAGGER-API.yaml
openapi: 3.0.0
info:
  title: Docs Api Monopatines
  description: Utilizacion de api de uso y gestion de app de monopatines
  version: "2.0"
servers:
- url: https://virtserver.swaggerhub.com/LEOPOLDOVILLAO14/tpe-arquitecturas/2.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8080
  description: Swagger
...
paths:
  /auth/admin:
    post:
      tags:
      - Auth
      summary: Crear una cuenta de rol admin.
      description: Registrar una cuenta con rol ADMIN. Ideal para acceder a todos los endpoint de la API.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPersonalAccountDTO'
        required: true
      responses:
        "201":
          $ref: '#/components/responses/AuthRegisterResponse'
  /auth/mantenimiento:
    post:
      tags:
      - Auth
      summary: Crear una cuenta de rol mantenimiento.
      description: Registrar una cuenta con rol MANTENIMIENTO. Ideal para marcar monopatines como "en mantenimiento" y acceder a reportes del area de mantenimiento.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPersonalAccountDTO'
        required: true
      responses:
        "201":
          $ref: '#/components/responses/AuthRegisterResponse'
  /auth/cliente:
    post:
      tags:
      - Auth
      summary: Crear una cuenta de rol cliente.
      description: "Registrar una cuenta con rol CLIENTE. Es el rol mas restringido, se pueden crear viajes, crear usuarios, recargar saldo de cuentas y ver paradas con monopatines disponibles cercanas."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewClientAccountDTO'
        required: true
      responses:
        "201":
          $ref: '#/components/responses/AuthRegisterResponse'
  /auth/login:
    post:
      tags:
      - Auth
      summary: Autenticarse con una cuenta ya registrada.
      description: Proporciona un token JWT que otorga permisos de acuerdo al Rol de la cuenta con la que se autentica. Este token debe ser utilizado en todos los endpoints siguientes.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDTO'
        required: true
      responses:
        "200":
          $ref: '#/components/responses/AuthSuccessResponse'
...
  /cuentas/{id}:
    get:
      tags:
      - Cuentas
      summary: "Obtiene la cuenta asociada al {id}."
      description: Ideal para ver detalles especificos de una cuenta en particular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la cuenta a buscar.
      responses:
        "200":
          description: Cuenta obtenida con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cuenta'
    delete:
      tags:
      - Cuentas
      summary: "Elimina la cuenta asociada al {id}."
      description: Ideal para eliminar una cuenta.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la cuenta a eliminar.
      responses:
        "200":
          description: Cuenta eliminada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAccountDTO'
  /cuentas/inhabilitar/{id}:
    put:
      tags:
      - Cuentas
      summary: "Inhabilita para hacer viajes a la cuenta asociada al {id}."
      description: Ideal para no permitir viajar a cuentas con movimientos sospechosos.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la cuenta a deshabilitar.
      responses:
        "200":
          description: Cuenta deshabilitada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisableAccountDTO'
  /cuentas/recargar/{id}:
    put:
      tags:
      - Cuentas
      summary: "Recarga saldo a la cuenta asociada al {id}."
      description: Ideal para cargar salgo y poder hacer viajes.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la cuenta a acreditar.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RechargeDTO'
        required: true
      responses:
        "200":
          description: Cuenta recargada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRechargeDTO'
...
  /paradas/:
    get:
      tags:
      - Paradas
      summary: Obtiene la lista de todas las paradas creadas.
      description: Util para listar todas las paradas que hay al momento.
      responses:
        "200":
          description: Paradas obtenidas con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StopDTO'
    post:
      tags:
      - Paradas
      summary: Crea una nueva parada.
      description: Util para crear paradas para poder hacer viajes y asignarles monopatines.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisibleDataStopDTO'
        required: true
      responses:
        "201":
          description: Parada creada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisibleDataStopDTO'
  /paradas/{id}:
    get:
      tags:
      - Paradas
      summary: "Obtiene una parada asociada a su {id}."
      description: Util para ver detalles de una parada en particular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la parada a buscar.
      responses:
        "200":
          description: Parada obtenida con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopDTO'
    put:
      tags:
      - Paradas
      summary: "Modifica una parada asociada a su {id}."
      description: Util para cambiar algun atributo de una parada en particular.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la parada a modificar.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisibleDataStopDTO'
        required: true
      responses:
        "200":
          description: Parada modificada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisibleDataStopDTO'
    delete:
      tags:
      - Paradas
      summary: "Elimina una parada asociada a su {id}."
      description: Util para eliminar una parada.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la parada a eliminar.
      responses:
        "200":
          description: Parada eliminada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisibleDataStopDTO'
  /paradas/cercanas:
    get:
      tags:
      - Paradas
      summary: Obtiene una lista de paradas cercanas a una ubicacion de X e Y.
      description: "Util para obtener las paradas cercanas a una ubicacion (coordenadas X e Y), retorna una lista ordenada con las mas cercanas a X e Y mostradas primero. Ademas incluye la cantidad de monopatines disponibles en cada parada."
      parameters:
      - name: x
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Ubicacion en X de un usuario.
      - name: "y"
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: number
          description: Ubicacion en Y de un usuario.
      responses:
        "200":
          description: Lista de paradas cercanas obtenida con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopAvailabilityDTO'
...
  /tarifas/:
    get:
      tags:
      - Tarifas
      summary: Obtiene la lista de todas las tarifas creadas.
      description: Util para enlistar las tarifas que han sido creadas y ver fechas de vigencia de todas.
      responses:
        "200":
          description: Lista de tarifas obtenida con exito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tarifa'
    post:
      tags:
      - Tarifas
      summary: Crea una nueva tarifa.
      description: "Crea una tarifa para el cobro de los viajes, posee una fecha de vigencia la cual al momento de crear un viaje, tomará la fecha anterior mas cercana a crear el viaje y será la que se utilice para el cobro del viaje."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeDTO'
        required: true
      responses:
        "201":
          description: Tarifa creada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeDTO'
  /tarifas/{id}:
    put:
      tags:
      - Tarifas
      summary: "Modifica una tarifa asociada a su {id}."
      description: Util para modificar algun atributo de una tarifa. No recomendable cambiar la fecha de vigencia ya que podria afectar el funcionamiento del sistema.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la tarifa a modificar.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeDTO'
        required: true
      responses:
        "200":
          description: Tarifa modificada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeDTO'
    delete:
      tags:
      - Tarifas
      summary: "Elimina una tarifa asociada a su {id}."
      description: Util para eliminar una tarifa.
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          description: Id de la tarifa a eliminar.
      responses:
        "200":
          description: Tarifa eliminada con exito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarifa'
components:
  schemas:
    NewPersonalAccountDTO:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          type: string
          example: pablo128
        email:
          type: string
          example: pablo@gmail.com
        password:
          type: string
          example: thepassword2024
    NewClientAccountDTO:
      required:
      - email
      - id_mp
      - password
      - username
      type: object
      properties:
        username:
          type: string
          example: juanfer77
        email:
          type: string
          example: juanfer@gmail.com
        password:
          type: string
          example: themostsecuritypass777
        id_mp:
          type: string
          example: 9fe2c4e9-3bd6-4b43-9cdb-7e40e8b2a8b6
    ResponseAccountDTO:
      required:
      - email
      - username
      type: object
      properties:
        username:
          type: string
          example: juanfer77
        email:
          type: string
          example: juanfer@gmail.com
    LoginDTO:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
          example: juanfer77
        email:
          type: string
          example: juanfer@gmail.com
    Rol:
      type: object
      properties:
        id_rol:
          type: number
          example: 2
        type:
          type: string
          example: cliente
    Cuenta:
      type: object
      properties:
        id_account:
          type: number
          example: 45
        role:
          $ref: '#/components/schemas/Rol'
        username:
          type: string
          example: pepito2024
        email:
          type: string
          example: pepito@gmail.com
        password:
          type: string
          example: UIQWYEUQP?;2873WOEER%6$QIUWE
        dischargeDate:
          type: string
          example: 2024-11-01
        isAvailable:
          type: boolean
...
    DisableAccountDTO:
      type: object
      properties:
        username:
          type: string
          example: juancito18
        email:
          type: string
          example: juancitoperez1818@gmail.com
        isAvailable:
          type: boolean
          example: false
    RechargeDTO:
      required:
      - amount
      type: object
      properties:
        amount:
          type: number
          example: 50000
    ResponseRechargeDTO:
      type: object
      properties:
        account:
          type: number
          example: 45
        newAmount:
          type: number
          example: 51250
...
    Point:
      type: object
      properties:
        x:
          type: number
          example: 5
        "y":
          type: number
          example: 5
...
    StopDTO:
      type: object
      properties:
        id_stop:
          type: number
          example: 2
        name:
          type: string
          example: Parada zona centro
        adress:
          type: string
          example: Avenida libertador 2018
        location:
          $ref: '#/components/schemas/Point'
...
    VisibleDataStopDTO:
      required:
      - adress
      - location
      - name
      type: object
      properties:
        name:
          type: string
          example: Parada zona centro
        adress:
          type: string
          example: Avenida libertador 1818
        location:
          $ref: '#/components/schemas/Point'
    StopAvailabilityDTO:
      type: object
      properties:
        name:
          type: string
          example: Parada zona campus universitario
        adress:
          type: string
          example: Reforma universitaria 2020
        location:
          $ref: '#/components/schemas/Point'
        quantityMonoAvailable:
          type: number
          example: 12
...
    Tarifa:
      type: object
      properties:
        id_fee:
          type: number
          example: 3
        normalFee:
          type: number
          example: 500
        additionalFee:
          type: number
          example: 1500
        dateValidity:
          type: string
          example: 2024-10-01
...
    FeeDTO:
      required:
      - additionalFee
      - dateValidity
      - normalFee
      type: object
      properties:
        normalFee:
          type: number
          example: 700
        additionalFee:
          type: number
          example: 1000
        dateValidity:
          type: string
          example: 2024-11-17
...
  responses:
    AuthRegisterResponse:
      description: CREATED
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResponseAccountDTO'
...
    AuthSuccessResponse:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              token:
                type: string
                example: JKHGSDYEGqywtewyEFYHAV625GE26Gqwbvq?=eIU26GHDKVG;EDUIOF
              type:
                type: string
                example: Bearer